backend/pyproject.toml
[tool.poetry]
name = "test-for-itfox"
version = "0.1.0"
description = ""
authors = ["Imwisagist@yandex.ru"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.2"
djangorestframework = "^3.14.0"
gunicorn = "^20.1.0"
psycopg2-binary = "^2.9.6"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
